requirements.txt
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
